+++ test/functional_test.dart
@@ -30,9 +30,18 @@ Future<void> main() async {
         await proc.stderr
             .transform(utf8.decoder)
             .transform(const LineSplitter())
-            .take(1)
-            .join(),
-        'Total coverage of 91% is below expected 100%.');
+            // .take(1)
+            .join('\n'),
+        equals([
+          'Total coverage of 91% is below expected 100%.',
+          'Top uncovered files:',
+          '/home/f3ath/project/marker/lib/flavors.dart',
+          'Lines (12): 36-38, 40-42, 45-49, 51',
+          '/home/f3ath/project/marker/lib/ast.dart',
+          'Lines (1): 116',
+          '/home/f3ath/project/marker/lib/src/flavors/changelog.dart',
+          'Lines (1): 10',
+        ].join('\n')));
   });
 
   test('Coverage not found', () async {