+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.5] - 2022-09-08
+### Changed
+- Rely on lcov\_tracefile for parsing
+- Group uncovered lines into ranges
+- Bump dependencies
+
 ## [0.0.4] - 2022-02-28
 ### Changed
 - Relaxed min Dart version to 2.15
@@ -14,6 +20,7 @@
 ### Added
 - Initial version
 
+[0.0.5]: https://github.com/f3ath/check-coverage/compare/0.0.4...0.0.5
 [0.0.4]: https://github.com/f3ath/check-coverage/compare/0.0.3...0.0.4
 [0.0.3]: https://github.com/f3ath/check-coverage/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/f3ath/check-coverage/compare/0.0.1...0.0.2